// sourcecode/BTrak.Packages/snova-dashboard-module/Dashboard-module-components/dashboard-components/src/lib/snovasys-dashboard-management/models/project-feature.model.ts
export class ProjectFeature {
    projectFeatureId: string;
    projectFeatureName: string;
    projectId: string;
    projectFeatureResponsiblePersonId: string;
    isArchived: boolean;
    IsDelete: boolean;
    projectFeatureResponsiblePersonName: string;
    profileImage: string;
    roleName: string;
    projectName: string;
    projectFeatureResponsiblePerson: any;
    timeStamp: any;
    createdDateTime: any;
}